feat: 区块加载

+++ service-game/src/main/java/com/t13max/game/world/chunk/ChunkGenerator.java
@@ -8,7 +8,7 @@ import com.t13max.persist.data.chunk.ChunkData;
  */
 public abstract class ChunkGenerator {
 
-    public abstract ChunkData generateChunkData(long chunkId);
+    public abstract ChunkData generateChunkData(int chunkId);
 
 
 }
